+++ wsgidav/default_conf.py
@@ -34,10 +34,14 @@ DEFAULT_CONFIG = {
     "mount_path": None,  # Application root, e.g. <mount_path>/<share_name>/<res_path>
     "provider_mapping": {},
     "add_header_MS_Author_Via": True,
-    "unquote_path_info": False,  # See issue #8
-    "re_encode_path_info": None,  # (See issue #73) None: activate on Python 3
+    "hotfixes": {
+        "emulate_win32_lastmod": False,  # True: support Win32LastModifiedTime
+        "re_encode_path_info": None,  # (See issue #73) None: activate on Python 3
+        "unquote_path_info": False,  # See issue #8
+        "win_accept_anonymous_options": False,
+        "winxp_accept_root_share_login": False,  # Was True in v2.4
+    },
     "property_manager": None,  # True: use property_manager.PropertyManager
-    "emulate_win32_lastmod": False,  # True: support Win32LastModifiedTime
     "mutable_live_props": [],
     "lock_manager": True,  # True: use lock_manager.LockManager
     "middleware_stack": [
@@ -72,7 +76,7 @@ DEFAULT_CONFIG = {
     "error_printer": {"catch_all": True},  # False,
     "enable_loggers": [],
     "dir_browser": {
-        # "enable": True,  # Render HTML listing for GET requests on collections
+        "enable": True,  # Render HTML listing for GET requests on collections
         # List of fnmatch patterns:
         "ignore": [
             ".DS_Store",  # macOS folder meta data