+++ wsgidav/middleware.py
@@ -33,3 +33,6 @@ class BaseMiddleware(object):
 
     def __str__(self):
         return "{}.{}".format(self.__module__, self.__class__.__name__)
+
+    def is_disabled(self):
+        return False